Refactored EditorContent component
