+++ lib/components/EditorContent/highlight.js
