+++ lib/components/EditorContent/index.jsx
@@ -3,8 +3,8 @@ import React from "react";
 import classnames from "classnames";
 import DOMPurify from "dompurify";
 
-import { EDITOR_CONTENT_CLASSNAME } from "constants/common";
-import highlightCode from "utils/highlightCode";
+import { EDITOR_CONTENT_CLASSNAME, SANITIZE_OPTIONS } from "./constants";
+import { highlightCode } from "./utils";
 
 const EditorContent = ({ content = "", className, ...otherProps }) => {
   const htmlContent = highlightCode(content);
@@ -16,10 +16,7 @@ const EditorContent = ({ content = "", className, ...otherProps }) => {
         [className]: className,
       })}
       dangerouslySetInnerHTML={{
-        __html: sanitize(htmlContent, {
-          ADD_TAGS: ["iframe"],
-          ADD_ATTR: ["target", "allow", "allowfullscreen", "frameborder"],
-        }),
+        __html: sanitize(htmlContent, SANITIZE_OPTIONS),
       }}
       data-cy="neeto-editor-content"
       {...otherProps}
