+++ lib/components/EditorContent/utils.js
@@ -3,7 +3,7 @@ import React from "react";
 import { lowlight } from "lowlight";
 import { renderToString } from "react-dom/server";
 
-import { CODE_BLOCK_REGEX } from "constants/common";
+import { CODE_BLOCK_REGEX } from "./constants";
 
 const buildReactElementFromAST = element => {
   if (element.tagName) {
@@ -16,7 +16,7 @@ const buildReactElementFromAST = element => {
   return element.value;
 };
 
-const highlightCode = content =>
+export const highlightCode = content =>
   content.replace(CODE_BLOCK_REGEX, (_, code) => {
     const highlightedAST = lowlight.highlightAuto(
       code.replace(/&gt;/g, ">").replace(/&lt;/g, "<").replace(/&amp;/g, "&")
@@ -26,5 +26,3 @@ const highlightCode = content =>
     );
     return `<pre><code>${renderToString(highlightedNode)}</code></pre>`;
   });
-
-export default highlightCode;
