+++ lib/constants/common.js
@@ -1,5 +1,3 @@
-export const EDITOR_CONTENT_CLASSNAME = "neeto-editor-content";
-
 export const EDITOR_PADDING_SIZE = 12;
 
 export const EDITOR_BORDER_SIZE = 1;
@@ -41,5 +39,3 @@ export const FONT_SIZE_OPTIONS = [
 ];
 
 export const DIRECT_UPLOAD_ENDPOINT = "/api/direct_uploads";
-
-export const CODE_BLOCK_REGEX = /<pre><code>([\S\s]*?)<\/code><\/pre>/gim;
