+++ lib/index.js
@@ -1,8 +1,5 @@
-// eslint-disable-next-line import/order
+import Editor from "./components/Editor";
+import EditorContent from "./components/EditorContent";
 import "./index.scss";
 
-import Editor from "./components/Editor";
-import EditorContent from "./components/Editor/EditorContent";
-import { EDITOR_OPTIONS } from "./constants/common";
-
-export { Editor, EditorContent, EDITOR_OPTIONS };
+export { Editor, EditorContent };
